addrstrings: fixed name typo, added minimal implementation and documentation.
